Clean up atomic action tutorials
