+++ scripts/tutorials/atomic_action/move_held_object.py
@@ -14,14 +14,13 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-"""Demonstrate Place after a PickUp precondition has created held-object state."""
+"""Demonstrate moving a held object to an object-centric target pose."""
 
 from __future__ import annotations
 
 import argparse
 import sys
 import time
-from collections.abc import Sequence
 from pathlib import Path
 
 _REPO_ROOT = Path(__file__).resolve().parents[3]
@@ -36,18 +35,20 @@ from embodichain.lab.sim import SimulationManager, SimulationManagerCfg
 from embodichain.lab.sim.atomic_actions import (
     AntipodalAffordance,
     AtomicActionEngine,
-    EndEffectorPoseTarget,
     GraspTarget,
+    HeldObjectPoseTarget,
+    MoveEndEffector,
+    MoveEndEffectorCfg,
+    MoveHeldObject,
+    MoveHeldObjectCfg,
     ObjectSemantics,
     PickUp,
     PickUpCfg,
-    Place,
-    PlaceCfg,
+    EndEffectorPoseTarget,
 )
 from embodichain.lab.sim.cfg import (
     JointDrivePropertiesCfg,
     LightCfg,
-    MarkerCfg,
     RenderCfg,
     RigidBodyAttributesCfg,
     RigidObjectCfg,
@@ -67,6 +68,12 @@ from embodichain.toolkits.graspkit.pg_grasp.gripper_collision_checker import (
 )
 from embodichain.utils import logger
 from embodichain.utils.math import matrix_from_euler
+from scripts.tutorials.atomic_action.tutorial_utils import (
+    draw_axis_marker,
+    get_tutorial_window_size,
+    start_auto_play_recording,
+    stop_auto_play_recording,
+)
 
 GRIPPER_URDF_PATH = "DH_PGI_140_80/DH_PGI_140_80.urdf"
 GRIPPER_HAND_JOINT_PATTERN = "GRIPPER_FINGER1_JOINT_1"
@@ -79,38 +86,26 @@ GRIPPER_TCP_Z = 0.15
 OBJECT_LABEL = "sugar_box"
 OBJECT_MESH_PATH = "SugarBox/sugar_box_usd/sugar_box.usda"
 OBJECT_XY = (-0.42, -0.08)
-OBJECT_MIN_HAND_CLOSE_QPOS = 0.024
+OBJECT_CLEARANCE = 0.0
 OBJECT_APPROACH_DIRECTION = (0.0, 0.0, -1.0)
+OBJECT_MIN_HAND_CLOSE_QPOS = 0.024
 OBJECT_INIT_ROT = (0.0, 0.0, 0.0)
 OBJECT_BODY_SCALE = (0.8, 0.8, 0.8)
 OBJECT_MASS = 0.05
 OBJECT_USE_USD_PROPERTIES = False
 
+MOVE_SAMPLE_INTERVAL = 60
 PICK_SAMPLE_INTERVAL = 120
-PLACE_SAMPLE_INTERVAL = 120
+MOVE_HELD_OBJECT_SAMPLE_INTERVAL = 120
 HAND_INTERP_STEPS = 12
 POST_TRAJECTORY_STEPS = 240
-EEF_AXIS_LEN = 0.06
-EEF_AXIS_SIZE = 0.003
-DEFAULT_AUTO_PLAY_LOOK_AT = (
-    (-1.6, -1.5, 1.2),
-    (0.0, 0.0, 0.25),
-    (0.0, 0.0, 1.0),
-)
-RECORD_WIDTH = 640
-RECORD_HEIGHT = 480
-VIEWER_WIDTH = 1600
-VIEWER_HEIGHT = 900
-AUTO_PLAY_RECORD_FPS = 20
-AUTO_PLAY_RECORD_MAX_MEMORY = 2048
 TABLE_SIZE = [1.0, 1.4, 0.05]
 TABLE_TOP_Z = -0.045
-PLACE_LIFT_HEIGHT = 0.14
 
 
 def parse_arguments() -> argparse.Namespace:
     parser = argparse.ArgumentParser(
-        description="Demonstrate Place by picking an object and releasing it at a target pose."
+        description="Demonstrate MoveHeldObject holding a sugar box in the gripper."
     )
     add_env_launcher_args_to_parser(parser)
     parser.add_argument(
@@ -135,57 +130,6 @@ def parse_arguments() -> argparse.Namespace:
         help="Do not draw the current end-effector/TCP coordinate frame before planning.",
     )
     return parser.parse_args()
-
-
-def get_tutorial_window_size(args: argparse.Namespace) -> tuple[int, int]:
-    """Return the viewer window size used by this tutorial."""
-    return VIEWER_WIDTH, VIEWER_HEIGHT
-
-
-def start_auto_play_recording(
-    sim: SimulationManager,
-    args: argparse.Namespace,
-    video_prefix: str,
-    look_at: tuple[
-        Sequence[float],
-        Sequence[float],
-        Sequence[float],
-    ] = DEFAULT_AUTO_PLAY_LOOK_AT,
-) -> bool:
-    """Start recording for ``--auto_play`` tutorial runs."""
-    if not getattr(args, "auto_play", False):
-        return False
-
-    original_width = sim.sim_config.width
-    original_height = sim.sim_config.height
-    try:
-        sim.sim_config.width = RECORD_WIDTH
-        sim.sim_config.height = RECORD_HEIGHT
-        if not sim.start_window_record(
-            fps=AUTO_PLAY_RECORD_FPS,
-            max_memory=AUTO_PLAY_RECORD_MAX_MEMORY,
-            video_prefix=video_prefix,
-            look_at=look_at,
-            use_sim_time=True,
-        ):
-            raise RuntimeError("Failed to start auto_play recording.")
-    finally:
-        sim.sim_config.width = original_width
-        sim.sim_config.height = original_height
-    return True
-
-
-def stop_auto_play_recording(
-    sim: SimulationManager,
-    recording_started: bool,
-) -> None:
-    """Stop recording and wait until the mp4 has been written."""
-    if not recording_started:
-        return
-
-    if sim.is_window_recording():
-        sim.stop_window_record()
-    sim.wait_window_record_saves()
 
 
 def initialize_simulation(args: argparse.Namespace) -> SimulationManager:
@@ -214,6 +158,7 @@ def initialize_simulation(args: argparse.Namespace) -> SimulationManager:
 def create_robot(sim: SimulationManager, position=(0.0, 0.0, 0.0)) -> Robot:
     ur5_urdf_path = get_data_path("UniversalRobots/UR5/UR5.urdf")
     gripper_urdf_path = get_data_path(GRIPPER_URDF_PATH)
+
     cfg = RobotCfg(
         uid="UR5",
         urdf_cfg=URDFCfg(
@@ -294,22 +239,8 @@ def compute_tabletop_init_pos(
     upright_rot = matrix_from_euler(rot, "XYZ")[0]
     rotated_vertices = vertices @ upright_rot.T
     bottom_z = rotated_vertices[:, 2].min().item()
-    return (OBJECT_XY[0], OBJECT_XY[1], TABLE_TOP_Z - bottom_z)
-
-
-def get_hand_open_close_qpos(
-    robot: Robot, device: torch.device
-) -> tuple[torch.Tensor, torch.Tensor]:
-    hand_limits = robot.get_qpos_limits(name="hand")[0].to(
-        device=device, dtype=torch.float32
-    )
-    hand_open = hand_limits[:, 0]
-    hand_close_limit = hand_limits[:, 1]
-    hand_close = torch.minimum(
-        hand_close_limit,
-        torch.full_like(hand_close_limit, OBJECT_MIN_HAND_CLOSE_QPOS),
-    )
-    return hand_open, hand_close
+    z = TABLE_TOP_Z + OBJECT_CLEARANCE - bottom_z
+    return (OBJECT_XY[0], OBJECT_XY[1], z)
 
 
 def settle_object(sim: SimulationManager, obj: RigidObject, step: int = 5) -> None:
@@ -364,66 +295,62 @@ def create_object_semantics(
     )
 
 
+def get_hand_open_close_qpos(
+    robot: Robot, device: torch.device
+) -> tuple[torch.Tensor, torch.Tensor]:
+    hand_limits = robot.get_qpos_limits(name="hand")[0].to(
+        device=device, dtype=torch.float32
+    )
+    hand_open = hand_limits[:, 0]
+    hand_close_limit = hand_limits[:, 1]
+    hand_close = torch.minimum(
+        hand_close_limit,
+        torch.full_like(hand_close_limit, OBJECT_MIN_HAND_CLOSE_QPOS),
+    )
+    return hand_open, hand_close
+
+
 def make_pre_pick_eef_pose(robot: Robot, position: torch.Tensor) -> torch.Tensor:
     pose = robot.compute_fk(
         qpos=robot.get_qpos(name="arm"),
         name="arm",
         to_matrix=True,
-    ).clone()
-    pose[:, :3, 3] = position
+    )[0].clone()
+    pose[:3, 3] = position
     return pose
 
 
-def initialize_pre_pick_robot_pose(
-    robot: Robot,
-    obj: RigidObject,
-    hand_open: torch.Tensor,
-) -> None:
-    obj_pose = obj.get_local_pose(to_matrix=True)
-    move_position = obj_pose[:, :3, 3].clone()
-    move_position[:, 2] = 0.36
-    pre_pick_pose = make_pre_pick_eef_pose(robot, move_position)
-    ik_success, arm_qpos = robot.compute_ik(
-        pose=pre_pick_pose,
-        joint_seed=robot.get_qpos(name="arm"),
-        name="arm",
-    )
-    if not torch.all(ik_success):
-        raise RuntimeError("Failed to initialize the robot at the pre-pick pose.")
-
-    n_envs = robot.get_qpos().shape[0]
-    hand_qpos = hand_open.unsqueeze(0).repeat(n_envs, 1)
-    for target in (False, True):
-        robot.set_qpos(arm_qpos, name="arm", target=target)
-        robot.set_qpos(hand_qpos, name="hand", target=target)
-    robot.clear_dynamics()
-
-
-def make_place_eef_pose(device: torch.device) -> torch.Tensor:
+def make_object_target_pose(device: torch.device) -> torch.Tensor:
     pose = torch.eye(4, dtype=torch.float32, device=device)
     pose[:3, :3] = torch.tensor(
         [
-            [-0.0539, -0.9985, -0.0022],
-            [-0.9977, 0.0540, -0.0401],
-            [0.0401, 0.0000, -0.9992],
+            [0.0, 0.0, -1.0],
+            [0.0, 1.0, 0.0],
+            [1.0, 0.0, 0.0],
         ],
         dtype=torch.float32,
         device=device,
     )
-    pose[:3, 3] = torch.tensor([-0.20, 0.28, 0.1], dtype=torch.float32, device=device)
+    pose[:3, 3] = torch.tensor([-0.3, -0.3, 0.5], dtype=torch.float32, device=device)
     return pose
 
 
 def compute_pick_close_end_step() -> int:
     motion_waypoints = PICK_SAMPLE_INTERVAL - HAND_INTERP_STEPS
     n_approach = int(round(motion_waypoints) * 0.6)
-    return n_approach + HAND_INTERP_STEPS
+    return MOVE_SAMPLE_INTERVAL + n_approach + HAND_INTERP_STEPS
 
 
-def build_pickup_cfg(
-    hand_open: torch.Tensor, hand_close: torch.Tensor, device: torch.device
-) -> PickUpCfg:
-    return PickUpCfg(
+def build_action_sequence(
+    hand_open: torch.Tensor,
+    hand_close: torch.Tensor,
+    device: torch.device,
+) -> list:
+    move_cfg = MoveEndEffectorCfg(
+        control_part="arm",
+        sample_interval=MOVE_SAMPLE_INTERVAL,
+    )
+    pickup_cfg = PickUpCfg(
         control_part="arm",
         hand_control_part="hand",
         hand_open_qpos=hand_open,
@@ -436,38 +363,16 @@ def build_pickup_cfg(
         sample_interval=PICK_SAMPLE_INTERVAL,
         hand_interp_steps=HAND_INTERP_STEPS,
     )
-
-
-def build_place_cfg(hand_open: torch.Tensor, hand_close: torch.Tensor) -> PlaceCfg:
-    return PlaceCfg(
+    move_held_object_cfg = MoveHeldObjectCfg(
         control_part="arm",
         hand_control_part="hand",
-        hand_open_qpos=hand_open,
         hand_close_qpos=hand_close,
-        lift_height=PLACE_LIFT_HEIGHT,
-        sample_interval=PLACE_SAMPLE_INTERVAL,
-        hand_interp_steps=HAND_INTERP_STEPS,
+        sample_interval=MOVE_HELD_OBJECT_SAMPLE_INTERVAL,
     )
+    return [move_cfg, pickup_cfg, move_held_object_cfg]
 
 
-def draw_current_eef_axis(sim: SimulationManager, robot: Robot) -> None:
-    eef_pose = robot.compute_fk(
-        qpos=robot.get_qpos(name="arm"),
-        name="arm",
-        to_matrix=True,
-    )
-    sim.draw_marker(
-        cfg=MarkerCfg(
-            name="current_eef_axis",
-            marker_type="axis",
-            axis_xpos=eef_pose,
-            axis_size=EEF_AXIS_SIZE,
-            axis_len=EEF_AXIS_LEN,
-        )
-    )
-
-
-def run_place_demo(args: argparse.Namespace) -> None:
+def run_move_held_object_demo(args: argparse.Namespace) -> None:
     sim = initialize_simulation(args)
     robot = create_robot(sim)
     create_table(sim)
@@ -479,39 +384,52 @@ def run_place_demo(args: argparse.Namespace) -> None:
         cfg=MotionGenCfg(planner_cfg=ToppraPlannerCfg(robot_uid=robot.uid))
     )
     hand_open, hand_close = get_hand_open_close_qpos(robot, sim.device)
-    initialize_pre_pick_robot_pose(robot, obj, hand_open)
-    pickup_cfg = build_pickup_cfg(hand_open, hand_close, sim.device)
-    place_cfg = build_place_cfg(hand_open, hand_close)
-    place_eef_pose = make_place_eef_pose(sim.device)
-
+    action_cfgs = build_action_sequence(hand_open, hand_close, sim.device)
     atomic_engine = AtomicActionEngine(motion_generator=motion_gen)
-    atomic_engine.register(PickUp(motion_gen, cfg=pickup_cfg))
-    atomic_engine.register(Place(motion_gen, cfg=place_cfg))
+    _action_classes = {
+        "move_end_effector": MoveEndEffector,
+        "pick_up": PickUp,
+        "move_held_object": MoveHeldObject,
+    }
+    for cfg in action_cfgs:
+        atomic_engine.register(_action_classes[cfg.name](motion_gen, cfg=cfg))
+
+    obj_pose = obj.get_local_pose(to_matrix=True)
+    move_position = obj_pose[0, :3, 3].clone()
+    move_position[2] = 0.36
+    move_target = make_pre_pick_eef_pose(robot, move_position)
+    object_target_pose = make_object_target_pose(sim.device)
+    move_held_object_target = HeldObjectPoseTarget(
+        object_target_pose=object_target_pose
+    )
 
     if not args.headless:
         sim.open_window()
     if not args.no_vis_eef_axis:
-        draw_current_eef_axis(sim, robot)
+        draw_axis_marker(sim, "move_held_object_target_axis", object_target_pose)
     if not args.auto_play:
-        input("Inspect the object, then press Enter to plan PickUp -> Place...")
+        input("Inspect the sugar box, then press Enter to plan...")
 
-    place_target = EndEffectorPoseTarget(xpos=place_eef_pose)
-    logger.log_info("Planning PickUp precondition -> Place release trajectory")
+    logger.log_info("Planning move_end_effector -> pick_up -> move_held_object")
+    start_time = time.time()
     is_success, traj, _ = atomic_engine.run(
         steps=[
+            ("move_end_effector", EndEffectorPoseTarget(xpos=move_target)),
             ("pick_up", GraspTarget(semantics=semantics)),
-            ("place", place_target),
+            ("move_held_object", move_held_object_target),
         ]
     )
+    cost_time = time.time() - start_time
+    logger.log_info(f"Plan trajectory cost time: {cost_time:.2f} seconds")
     if not is_success:
-        logger.log_warning("Failed to plan Place demo trajectory.")
+        logger.log_warning("Failed to plan move_held_object demo trajectory.")
         return
 
     if not args.auto_play:
-        input("Press Enter to replay the Place demo...")
+        input("Press Enter to replay the move_held_object demo...")
 
     recording_started = start_auto_play_recording(
-        sim, args, video_prefix="place_auto_play"
+        sim, args, video_prefix="move_held_object_auto_play"
     )
     try:
         post_grasp_clear_step = compute_pick_close_end_step()
@@ -525,7 +443,8 @@ def run_place_demo(args: argparse.Namespace) -> None:
                 logger.log_info(f"Object dynamics cleared after grasp at step={i}")
             time.sleep(1e-2)
 
-        logger.log_info("Place opens the gripper and clears WorldState.held_object.")
+        logger.log_info("MoveHeldObject keeps the sugar box suspended in the gripper.")
+
         final_qpos = traj[:, -1, :]
         for i in range(POST_TRAJECTORY_STEPS):
             robot.set_qpos(final_qpos)
@@ -540,7 +459,7 @@ def run_place_demo(args: argparse.Namespace) -> None:
 
 def main() -> None:
     args = parse_arguments()
-    run_place_demo(args)
+    run_move_held_object_demo(args)
 
 
 if __name__ == "__main__":
