+++ scripts/tutorials/atomic_action/move_joints.py
@@ -14,14 +14,13 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-"""Demonstrate MoveEndEffector with a single pose target."""
+"""Demonstrate MoveJoints with named and explicit joint-space targets."""
 
 from __future__ import annotations
 
 import argparse
 import sys
 import time
-from collections.abc import Sequence
 from pathlib import Path
 
 _REPO_ROOT = Path(__file__).resolve().parents[3]
@@ -35,14 +34,14 @@ from embodichain.lab.gym.utils.gym_utils import add_env_launcher_args_to_parser
 from embodichain.lab.sim import SimulationManager, SimulationManagerCfg
 from embodichain.lab.sim.atomic_actions import (
     AtomicActionEngine,
-    EndEffectorPoseTarget,
-    MoveEndEffector,
-    MoveEndEffectorCfg,
+    JointPositionTarget,
+    MoveJoints,
+    MoveJointsCfg,
+    NamedJointPositionTarget,
 )
 from embodichain.lab.sim.cfg import (
     JointDrivePropertiesCfg,
     LightCfg,
-    MarkerCfg,
     RenderCfg,
     RobotCfg,
     URDFCfg,
@@ -51,31 +50,24 @@ from embodichain.lab.sim.objects import Robot
 from embodichain.lab.sim.planners import MotionGenerator, MotionGenCfg, ToppraPlannerCfg
 from embodichain.lab.sim.solvers import PytorchSolverCfg
 from embodichain.utils import logger
+from scripts.tutorials.atomic_action.tutorial_utils import (
+    draw_axis_marker,
+    get_tutorial_window_size,
+    start_auto_play_recording,
+    stop_auto_play_recording,
+)
 
 GRIPPER_URDF_PATH = "DH_PGI_140_80/DH_PGI_140_80.urdf"
 GRIPPER_HAND_JOINT_PATTERN = "GRIPPER_FINGER1_JOINT_1"
 GRIPPER_TCP_Z = 0.15
 
-MOVE_SAMPLE_INTERVAL = 80
+MOVE_JOINTS_SAMPLE_INTERVAL = 80
 POST_TRAJECTORY_STEPS = 120
-DEFAULT_AUTO_PLAY_LOOK_AT = (
-    (-1.6, -1.5, 1.2),
-    (0.0, 0.0, 0.25),
-    (0.0, 0.0, 1.0),
-)
-RECORD_WIDTH = 640
-RECORD_HEIGHT = 480
-VIEWER_WIDTH = 1600
-VIEWER_HEIGHT = 900
-AUTO_PLAY_RECORD_FPS = 20
-AUTO_PLAY_RECORD_MAX_MEMORY = 2048
-EEF_AXIS_LEN = 0.06
-EEF_AXIS_SIZE = 0.003
 
 
 def parse_arguments() -> argparse.Namespace:
     parser = argparse.ArgumentParser(
-        description="Demonstrate MoveEndEffector with a top-down target pose."
+        description="Demonstrate MoveJoints with named and explicit qpos targets."
     )
     add_env_launcher_args_to_parser(parser)
     parser.add_argument(
@@ -89,57 +81,6 @@ def parse_arguments() -> argparse.Namespace:
         help="Do not draw the current end-effector/TCP coordinate frame before planning.",
     )
     return parser.parse_args()
-
-
-def get_tutorial_window_size(args: argparse.Namespace) -> tuple[int, int]:
-    """Return the viewer window size used by this tutorial."""
-    return VIEWER_WIDTH, VIEWER_HEIGHT
-
-
-def start_auto_play_recording(
-    sim: SimulationManager,
-    args: argparse.Namespace,
-    video_prefix: str,
-    look_at: tuple[
-        Sequence[float],
-        Sequence[float],
-        Sequence[float],
-    ] = DEFAULT_AUTO_PLAY_LOOK_AT,
-) -> bool:
-    """Start recording for ``--auto_play`` tutorial runs."""
-    if not getattr(args, "auto_play", False):
-        return False
-
-    original_width = sim.sim_config.width
-    original_height = sim.sim_config.height
-    try:
-        sim.sim_config.width = RECORD_WIDTH
-        sim.sim_config.height = RECORD_HEIGHT
-        if not sim.start_window_record(
-            fps=AUTO_PLAY_RECORD_FPS,
-            max_memory=AUTO_PLAY_RECORD_MAX_MEMORY,
-            video_prefix=video_prefix,
-            look_at=look_at,
-            use_sim_time=True,
-        ):
-            raise RuntimeError("Failed to start auto_play recording.")
-    finally:
-        sim.sim_config.width = original_width
-        sim.sim_config.height = original_height
-    return True
-
-
-def stop_auto_play_recording(
-    sim: SimulationManager,
-    recording_started: bool,
-) -> None:
-    """Stop recording and wait until the mp4 has been written."""
-    if not recording_started:
-        return
-
-    if sim.is_window_recording():
-        sim.stop_window_record()
-    sim.wait_window_record_saves()
 
 
 def initialize_simulation(args: argparse.Namespace) -> SimulationManager:
@@ -204,84 +145,62 @@ def create_robot(sim: SimulationManager, position=(0.0, 0.0, 0.0)) -> Robot:
     return sim.add_robot(cfg=cfg)
 
 
-def make_top_down_eef_pose(device: torch.device) -> torch.Tensor:
-    pose = torch.eye(4, dtype=torch.float32, device=device)
-    pose[:3, :3] = torch.tensor(
-        [
-            [-0.0539, -0.9985, -0.0022],
-            [-0.9977, 0.0540, -0.0401],
-            [0.0401, 0.0000, -0.9992],
-        ],
-        dtype=torch.float32,
-        device=device,
-    )
-    pose[:3, 3] = torch.tensor([0.30, -0.20, 0.36], dtype=torch.float32, device=device)
-    return pose
+def make_arm_qpos(values: list[float], device: torch.device) -> torch.Tensor:
+    return torch.tensor(values, dtype=torch.float32, device=device)
 
 
-def format_tensor(tensor: torch.Tensor) -> str:
-    rounded = (tensor.detach().cpu() * 10000.0).round() / 10000.0
-    return str(rounded.tolist())
-
-
-def draw_current_eef_axis(sim: SimulationManager, robot: Robot) -> None:
+def draw_start_eef_axis(sim: SimulationManager, robot: Robot) -> None:
     eef_pose = robot.compute_fk(
         qpos=robot.get_qpos(name="arm"),
         name="arm",
         to_matrix=True,
     )
-    sim.draw_marker(
-        cfg=MarkerCfg(
-            name="current_eef_axis",
-            marker_type="axis",
-            axis_xpos=eef_pose,
-            axis_size=EEF_AXIS_SIZE,
-            axis_len=EEF_AXIS_LEN,
-        )
-    )
+    draw_axis_marker(sim, "move_joints_start_eef_axis", eef_pose)
 
 
-def run_move_end_effector_demo(args: argparse.Namespace) -> None:
+def run_move_joints_demo(args: argparse.Namespace) -> None:
     sim = initialize_simulation(args)
     robot = create_robot(sim)
     motion_gen = MotionGenerator(
         cfg=MotionGenCfg(planner_cfg=ToppraPlannerCfg(robot_uid=robot.uid))
     )
 
-    atomic_engine = AtomicActionEngine(motion_generator=motion_gen)
-    atomic_engine.register(
-        MoveEndEffector(
-            motion_gen,
-            cfg=MoveEndEffectorCfg(
-                control_part="arm",
-                sample_interval=MOVE_SAMPLE_INTERVAL,
-            ),
-        )
+    ready_qpos = make_arm_qpos([0.35, -1.20, 1.30, -1.65, -1.57, 0.20], sim.device)
+    home_qpos = make_arm_qpos([0.0, -1.57, 1.57, -1.57, -1.57, 0.0], sim.device)
+    move_joints_cfg = MoveJointsCfg(
+        control_part="arm",
+        sample_interval=MOVE_JOINTS_SAMPLE_INTERVAL,
+        named_joint_positions={"ready": ready_qpos},
     )
+
+    atomic_engine = AtomicActionEngine(motion_generator=motion_gen)
+    atomic_engine.register(MoveJoints(motion_gen, cfg=move_joints_cfg))
 
     if not args.headless:
         sim.open_window()
     if not args.no_vis_eef_axis:
-        draw_current_eef_axis(sim, robot)
+        draw_start_eef_axis(sim, robot)
     if not args.auto_play:
-        input("Inspect the robot, then press Enter to plan MoveEndEffector...")
+        input("Inspect the robot, then press Enter to plan MoveJoints...")
 
-    target_pose = make_top_down_eef_pose(sim.device)
     logger.log_info(
-        f"Planning MoveEndEffector to xpos={format_tensor(target_pose[:3, 3])}"
+        "Planning MoveJoints: NamedJointPositionTarget('ready') -> explicit home qpos"
     )
     is_success, traj, _ = atomic_engine.run(
-        steps=[("move_end_effector", EndEffectorPoseTarget(xpos=target_pose))]
+        steps=[
+            ("move_joints", NamedJointPositionTarget(name="ready")),
+            ("move_joints", JointPositionTarget(qpos=home_qpos)),
+        ]
     )
     if not is_success:
-        logger.log_warning("Failed to plan MoveEndEffector demo trajectory.")
+        logger.log_warning("Failed to plan MoveJoints demo trajectory.")
         return
 
     if not args.auto_play:
-        input("Press Enter to replay the MoveEndEffector demo...")
+        input("Press Enter to replay the MoveJoints demo...")
 
     recording_started = start_auto_play_recording(
-        sim, args, video_prefix="move_end_effector_auto_play"
+        sim, args, video_prefix="move_joints_auto_play"
     )
     try:
         for i in range(traj.shape[1]):
@@ -303,7 +222,7 @@ def run_move_end_effector_demo(args: argparse.Namespace) -> None:
 
 def main() -> None:
     args = parse_arguments()
-    run_move_end_effector_demo(args)
+    run_move_joints_demo(args)
 
 
 if __name__ == "__main__":
