+++ scripts/tutorials/atomic_action/place.py
@@ -14,14 +14,13 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-"""Demonstrate PickUp on an upright object with configurable approach."""
+"""Demonstrate Place after a PickUp precondition has created held-object state."""
 
 from __future__ import annotations
 
 import argparse
 import sys
 import time
-from collections.abc import Sequence
 from pathlib import Path
 
 _REPO_ROOT = Path(__file__).resolve().parents[3]
@@ -36,15 +35,17 @@ from embodichain.lab.sim import SimulationManager, SimulationManagerCfg
 from embodichain.lab.sim.atomic_actions import (
     AntipodalAffordance,
     AtomicActionEngine,
+    EndEffectorPoseTarget,
     GraspTarget,
     ObjectSemantics,
     PickUp,
     PickUpCfg,
+    Place,
+    PlaceCfg,
 )
 from embodichain.lab.sim.cfg import (
     JointDrivePropertiesCfg,
     LightCfg,
-    MarkerCfg,
     RenderCfg,
     RigidBodyAttributesCfg,
     RigidObjectCfg,
@@ -64,6 +65,12 @@ from embodichain.toolkits.graspkit.pg_grasp.gripper_collision_checker import (
 )
 from embodichain.utils import logger
 from embodichain.utils.math import matrix_from_euler
+from scripts.tutorials.atomic_action.tutorial_utils import (
+    draw_axis_marker,
+    get_tutorial_window_size,
+    start_auto_play_recording,
+    stop_auto_play_recording,
+)
 
 GRIPPER_URDF_PATH = "DH_PGI_140_80/DH_PGI_140_80.urdf"
 GRIPPER_HAND_JOINT_PATTERN = "GRIPPER_FINGER1_JOINT_1"
@@ -73,58 +80,30 @@ GRIPPER_ROOT_Z_WIDTH = 0.096
 GRIPPER_Y_THICKNESS = 0.040
 GRIPPER_TCP_Z = 0.15
 
-OBJECT_MIN_HAND_CLOSE_QPOS = 0.024
+OBJECT_LABEL = "sugar_box"
+OBJECT_MESH_PATH = "SugarBox/sugar_box_usd/sugar_box.usda"
 OBJECT_XY = (-0.42, -0.08)
-OBJECT_CLEARANCE = 0.0
-
-OBJECT_PRESETS = {
-    "sugar_box": {
-        "label": "sugar_box",
-        "mesh_path": "SugarBox/sugar_box_usd/sugar_box.usda",
-        "init_rot": (0.0, 0.0, 0.0),
-        "body_scale": (0.8, 0.8, 0.8),
-        "mass": 0.05,
-        "use_usd_properties": False,
-    },
-}
+OBJECT_MIN_HAND_CLOSE_QPOS = 0.024
+OBJECT_APPROACH_DIRECTION = (0.0, 0.0, -1.0)
+OBJECT_INIT_ROT = (0.0, 0.0, 0.0)
+OBJECT_BODY_SCALE = (0.8, 0.8, 0.8)
+OBJECT_MASS = 0.05
+OBJECT_USE_USD_PROPERTIES = False
 
 PICK_SAMPLE_INTERVAL = 120
+PLACE_SAMPLE_INTERVAL = 120
 HAND_INTERP_STEPS = 12
 POST_TRAJECTORY_STEPS = 240
-DEFAULT_AUTO_PLAY_LOOK_AT = (
-    (-1.6, -1.5, 1.2),
-    (0.0, 0.0, 0.25),
-    (0.0, 0.0, 1.0),
-)
-RECORD_WIDTH = 640
-RECORD_HEIGHT = 480
-VIEWER_WIDTH = 1600
-VIEWER_HEIGHT = 900
-AUTO_PLAY_RECORD_FPS = 20
-AUTO_PLAY_RECORD_MAX_MEMORY = 2048
-EEF_AXIS_LEN = 0.06
-EEF_AXIS_SIZE = 0.003
 TABLE_SIZE = [1.0, 1.4, 0.05]
 TABLE_TOP_Z = -0.045
-
-APPROACH_DIRECTIONS = {
-    "top": (0.0, 0.0, -1.0),
-    "side": (0.0, 1.0, 0.0),
-    "side_y": (0.0, -1.0, 0.0),
-}
+PLACE_LIFT_HEIGHT = 0.14
 
 
 def parse_arguments() -> argparse.Namespace:
     parser = argparse.ArgumentParser(
-        description="Demonstrate PickUp on an upright object."
+        description="Demonstrate Place by picking an object and releasing it at a target pose."
     )
     add_env_launcher_args_to_parser(parser)
-    parser.add_argument(
-        "--object",
-        choices=sorted(OBJECT_PRESETS.keys()),
-        default="sugar_box",
-        help="Object preset to pick.",
-    )
     parser.add_argument(
         "--n_sample",
         type=int,
@@ -142,76 +121,11 @@ def parse_arguments() -> argparse.Namespace:
         help="Run the viewer demo without waiting for keyboard input.",
     )
     parser.add_argument(
-        "--approach",
-        choices=["top", "side", "side_y", "custom"],
-        default="top",
-        help="Pick approach direction preset.",
-    )
-    parser.add_argument(
-        "--custom_approach_direction",
-        type=float,
-        nargs=3,
-        default=None,
-        metavar=("X", "Y", "Z"),
-        help="World-frame approach direction used when --approach custom.",
-    )
-    parser.add_argument(
         "--no_vis_eef_axis",
         action="store_true",
         help="Do not draw the current end-effector/TCP coordinate frame before planning.",
     )
     return parser.parse_args()
-
-
-def get_tutorial_window_size(args: argparse.Namespace) -> tuple[int, int]:
-    """Return the viewer window size used by this tutorial."""
-    return VIEWER_WIDTH, VIEWER_HEIGHT
-
-
-def start_auto_play_recording(
-    sim: SimulationManager,
-    args: argparse.Namespace,
-    video_prefix: str,
-    look_at: tuple[
-        Sequence[float],
-        Sequence[float],
-        Sequence[float],
-    ] = DEFAULT_AUTO_PLAY_LOOK_AT,
-) -> bool:
-    """Start recording for ``--auto_play`` tutorial runs."""
-    if not getattr(args, "auto_play", False):
-        return False
-
-    original_width = sim.sim_config.width
-    original_height = sim.sim_config.height
-    try:
-        sim.sim_config.width = RECORD_WIDTH
-        sim.sim_config.height = RECORD_HEIGHT
-        if not sim.start_window_record(
-            fps=AUTO_PLAY_RECORD_FPS,
-            max_memory=AUTO_PLAY_RECORD_MAX_MEMORY,
-            video_prefix=video_prefix,
-            look_at=look_at,
-            use_sim_time=True,
-        ):
-            raise RuntimeError("Failed to start auto_play recording.")
-    finally:
-        sim.sim_config.width = original_width
-        sim.sim_config.height = original_height
-    return True
-
-
-def stop_auto_play_recording(
-    sim: SimulationManager,
-    recording_started: bool,
-) -> None:
-    """Stop recording and wait until the mp4 has been written."""
-    if not recording_started:
-        return
-
-    if sim.is_window_recording():
-        sim.stop_window_record()
-    sim.wait_window_record_saves()
 
 
 def initialize_simulation(args: argparse.Namespace) -> SimulationManager:
@@ -240,7 +154,6 @@ def initialize_simulation(args: argparse.Namespace) -> SimulationManager:
 def create_robot(sim: SimulationManager, position=(0.0, 0.0, 0.0)) -> Robot:
     ur5_urdf_path = get_data_path("UniversalRobots/UR5/UR5.urdf")
     gripper_urdf_path = get_data_path(GRIPPER_URDF_PATH)
-
     cfg = RobotCfg(
         uid="UR5",
         urdf_cfg=URDFCfg(
@@ -291,29 +204,28 @@ def create_table(sim: SimulationManager) -> RigidObject:
     return sim.add_rigid_object(cfg=cfg)
 
 
-def create_pick_object(sim: SimulationManager, object_name: str) -> RigidObject:
-    preset = OBJECT_PRESETS[object_name]
+def create_pick_object(sim: SimulationManager) -> RigidObject:
     cfg = RigidObjectCfg(
-        uid=preset["label"],
-        shape=MeshCfg(fpath=get_data_path(preset["mesh_path"])),
+        uid=OBJECT_LABEL,
+        shape=MeshCfg(fpath=get_data_path(OBJECT_MESH_PATH)),
         attrs=RigidBodyAttributesCfg(
-            mass=preset["mass"],
+            mass=OBJECT_MASS,
             dynamic_friction=0.97,
             static_friction=0.99,
         ),
         max_convex_hull_num=16,
         init_pos=[OBJECT_XY[0], OBJECT_XY[1], 0.0],
-        init_rot=preset["init_rot"],
-        body_scale=preset["body_scale"],
-        use_usd_properties=preset["use_usd_properties"],
+        init_rot=OBJECT_INIT_ROT,
+        body_scale=OBJECT_BODY_SCALE,
+        use_usd_properties=OBJECT_USE_USD_PROPERTIES,
     )
     obj = sim.add_rigid_object(cfg=cfg)
-    obj.cfg.init_pos = _compute_tabletop_init_pos(obj, cfg.init_rot)
+    obj.cfg.init_pos = compute_tabletop_init_pos(obj, cfg.init_rot)
     obj.reset()
     return obj
 
 
-def _compute_tabletop_init_pos(
+def compute_tabletop_init_pos(
     obj: RigidObject, init_rot: tuple[float, float, float]
 ) -> tuple[float, float, float]:
     vertices = obj.get_vertices(env_ids=[0], scale=True)[0]
@@ -322,8 +234,22 @@ def _compute_tabletop_init_pos(
     upright_rot = matrix_from_euler(rot, "XYZ")[0]
     rotated_vertices = vertices @ upright_rot.T
     bottom_z = rotated_vertices[:, 2].min().item()
-    z = TABLE_TOP_Z + OBJECT_CLEARANCE - bottom_z
-    return (OBJECT_XY[0], OBJECT_XY[1], z)
+    return (OBJECT_XY[0], OBJECT_XY[1], TABLE_TOP_Z - bottom_z)
+
+
+def get_hand_open_close_qpos(
+    robot: Robot, device: torch.device
+) -> tuple[torch.Tensor, torch.Tensor]:
+    hand_limits = robot.get_qpos_limits(name="hand")[0].to(
+        device=device, dtype=torch.float32
+    )
+    hand_open = hand_limits[:, 0]
+    hand_close_limit = hand_limits[:, 1]
+    hand_close = torch.minimum(
+        hand_close_limit,
+        torch.full_like(hand_close_limit, OBJECT_MIN_HAND_CLOSE_QPOS),
+    )
+    return hand_open, hand_close
 
 
 def settle_object(sim: SimulationManager, obj: RigidObject, step: int = 5) -> None:
@@ -361,9 +287,8 @@ def build_gripper_collision_cfg() -> GripperCollisionCfg:
 def create_object_semantics(
     obj: RigidObject, args: argparse.Namespace
 ) -> ObjectSemantics:
-    label = OBJECT_PRESETS[args.object]["label"]
     return ObjectSemantics(
-        label=label,
+        label=OBJECT_LABEL,
         geometry={
             "mesh_vertices": obj.get_vertices(env_ids=[0], scale=True)[0],
             "mesh_triangles": obj.get_triangles(env_ids=[0])[0],
@@ -377,40 +302,6 @@ def create_object_semantics(
         ),
         entity=obj,
     )
-
-
-def get_hand_open_close_qpos(
-    robot: Robot, device: torch.device
-) -> tuple[torch.Tensor, torch.Tensor]:
-    hand_limits = robot.get_qpos_limits(name="hand")[0].to(
-        device=device, dtype=torch.float32
-    )
-    hand_open = hand_limits[:, 0]
-    hand_close_limit = hand_limits[:, 1]
-    hand_close = torch.minimum(
-        hand_close_limit,
-        torch.full_like(hand_close_limit, OBJECT_MIN_HAND_CLOSE_QPOS),
-    )
-    return hand_open, hand_close
-
-
-def resolve_approach_direction(
-    args: argparse.Namespace, device: torch.device
-) -> torch.Tensor:
-    if args.approach == "custom":
-        if args.custom_approach_direction is None:
-            raise ValueError(
-                "--custom_approach_direction is required when --approach custom."
-            )
-        direction = args.custom_approach_direction
-    else:
-        direction = APPROACH_DIRECTIONS[args.approach]
-
-    approach_direction = torch.tensor(direction, dtype=torch.float32, device=device)
-    norm = torch.linalg.norm(approach_direction)
-    if norm < 1e-6:
-        raise ValueError("approach_direction must be non-zero.")
-    return approach_direction / norm
 
 
 def make_pre_pick_eef_pose(robot: Robot, position: torch.Tensor) -> torch.Tensor:
@@ -448,45 +339,38 @@ def initialize_pre_pick_robot_pose(
     robot.clear_dynamics()
 
 
+def make_place_eef_pose(device: torch.device) -> torch.Tensor:
+    pose = torch.eye(4, dtype=torch.float32, device=device)
+    pose[:3, :3] = torch.tensor(
+        [
+            [-0.0539, -0.9985, -0.0022],
+            [-0.9977, 0.0540, -0.0401],
+            [0.0401, 0.0000, -0.9992],
+        ],
+        dtype=torch.float32,
+        device=device,
+    )
+    pose[:3, 3] = torch.tensor([-0.20, 0.28, 0.1], dtype=torch.float32, device=device)
+    return pose
+
+
 def compute_pick_close_end_step() -> int:
     motion_waypoints = PICK_SAMPLE_INTERVAL - HAND_INTERP_STEPS
     n_approach = int(round(motion_waypoints) * 0.6)
     return n_approach + HAND_INTERP_STEPS
 
 
-def format_tensor(tensor: torch.Tensor) -> str:
-    rounded = (tensor.detach().cpu() * 10000.0).round() / 10000.0
-    return str(rounded.tolist())
-
-
-def draw_current_eef_axis(sim: SimulationManager, robot: Robot) -> None:
-    eef_pose = robot.compute_fk(
-        qpos=robot.get_qpos(name="arm"),
-        name="arm",
-        to_matrix=True,
-    )
-    sim.draw_marker(
-        cfg=MarkerCfg(
-            name="current_eef_axis",
-            marker_type="axis",
-            axis_xpos=eef_pose,
-            axis_size=EEF_AXIS_SIZE,
-            axis_len=EEF_AXIS_LEN,
-        )
-    )
-
-
 def build_pickup_cfg(
-    hand_open: torch.Tensor,
-    hand_close: torch.Tensor,
-    approach_direction: torch.Tensor,
+    hand_open: torch.Tensor, hand_close: torch.Tensor, device: torch.device
 ) -> PickUpCfg:
     return PickUpCfg(
         control_part="arm",
         hand_control_part="hand",
         hand_open_qpos=hand_open,
         hand_close_qpos=hand_close,
-        approach_direction=approach_direction,
+        approach_direction=torch.tensor(
+            OBJECT_APPROACH_DIRECTION, dtype=torch.float32, device=device
+        ),
         pre_grasp_distance=0.15,
         lift_height=0.16,
         sample_interval=PICK_SAMPLE_INTERVAL,
@@ -494,11 +378,23 @@ def build_pickup_cfg(
     )
 
 
-def run_pickup_demo(args: argparse.Namespace) -> None:
+def build_place_cfg(hand_open: torch.Tensor, hand_close: torch.Tensor) -> PlaceCfg:
+    return PlaceCfg(
+        control_part="arm",
+        hand_control_part="hand",
+        hand_open_qpos=hand_open,
+        hand_close_qpos=hand_close,
+        lift_height=PLACE_LIFT_HEIGHT,
+        sample_interval=PLACE_SAMPLE_INTERVAL,
+        hand_interp_steps=HAND_INTERP_STEPS,
+    )
+
+
+def run_place_demo(args: argparse.Namespace) -> None:
     sim = initialize_simulation(args)
     robot = create_robot(sim)
     create_table(sim)
-    obj = create_pick_object(sim, args.object)
+    obj = create_pick_object(sim)
 
     settle_object(sim, obj, step=5)
     semantics = create_object_semantics(obj, args)
@@ -506,38 +402,39 @@ def run_pickup_demo(args: argparse.Namespace) -> None:
         cfg=MotionGenCfg(planner_cfg=ToppraPlannerCfg(robot_uid=robot.uid))
     )
     hand_open, hand_close = get_hand_open_close_qpos(robot, sim.device)
-    approach_direction = resolve_approach_direction(args, sim.device)
     initialize_pre_pick_robot_pose(robot, obj, hand_open)
-    pickup_cfg = build_pickup_cfg(hand_open, hand_close, approach_direction)
+    pickup_cfg = build_pickup_cfg(hand_open, hand_close, sim.device)
+    place_cfg = build_place_cfg(hand_open, hand_close)
+    place_eef_pose = make_place_eef_pose(sim.device)
+
     atomic_engine = AtomicActionEngine(motion_generator=motion_gen)
     atomic_engine.register(PickUp(motion_gen, cfg=pickup_cfg))
+    atomic_engine.register(Place(motion_gen, cfg=place_cfg))
 
     if not args.headless:
         sim.open_window()
     if not args.no_vis_eef_axis:
-        draw_current_eef_axis(sim, robot)
+        draw_axis_marker(sim, "place_target_axis", place_eef_pose)
     if not args.auto_play:
-        input(f"Inspect the upright {args.object}, then press Enter to plan...")
+        input("Inspect the object, then press Enter to plan PickUp -> Place...")
 
-    logger.log_info(
-        f"Planning pick_up for {args.object} with "
-        f"approach_direction={format_tensor(approach_direction)}"
-    )
-    start_time = time.time()
+    place_target = EndEffectorPoseTarget(xpos=place_eef_pose)
+    logger.log_info("Planning PickUp precondition -> Place release trajectory")
     is_success, traj, _ = atomic_engine.run(
-        steps=[("pick_up", GraspTarget(semantics=semantics))]
+        steps=[
+            ("pick_up", GraspTarget(semantics=semantics)),
+            ("place", place_target),
+        ]
     )
-    cost_time = time.time() - start_time
-    logger.log_info(f"Plan trajectory cost time: {cost_time:.2f} seconds")
     if not is_success:
-        logger.log_warning("Failed to plan pickup demo trajectory.")
+        logger.log_warning("Failed to plan Place demo trajectory.")
         return
 
     if not args.auto_play:
-        input("Press Enter to replay the pickup demo...")
+        input("Press Enter to replay the Place demo...")
 
     recording_started = start_auto_play_recording(
-        sim, args, video_prefix=f"pickup_{args.object}_auto_play"
+        sim, args, video_prefix="place_auto_play"
     )
     try:
         post_grasp_clear_step = compute_pick_close_end_step()
@@ -551,10 +448,7 @@ def run_pickup_demo(args: argparse.Namespace) -> None:
                 logger.log_info(f"Object dynamics cleared after grasp at step={i}")
             time.sleep(1e-2)
 
-        logger.log_info(
-            f"PickUp keeps the upright {args.object} suspended in the gripper."
-        )
-
+        logger.log_info("Place opens the gripper and clears WorldState.held_object.")
         final_qpos = traj[:, -1, :]
         for i in range(POST_TRAJECTORY_STEPS):
             robot.set_qpos(final_qpos)
@@ -569,7 +463,7 @@ def run_pickup_demo(args: argparse.Namespace) -> None:
 
 def main() -> None:
     args = parse_arguments()
-    run_pickup_demo(args)
+    run_place_demo(args)
 
 
 if __name__ == "__main__":
